Refactor utils CMakeLists.txt file
Related-issue: APPLINK-21604
