+++ CMakeLists.txt
@@ -367,7 +367,7 @@ endmacro(GenerateInterface)
 macro(create_test NAME SOURCES LIBS)
   add_executable("${NAME}" ${CMAKE_SOURCE_DIR}/src/components/test_main.cc ${SOURCES})
   target_link_libraries("${NAME}" ${LIBS})
-  target_link_libraries("${NAME}" Utils)
+  target_link_libraries("${NAME}" utils)
   if(CMAKE_SYSTEM_NAME STREQUAL "QNX")
     add_test(${NAME} ${CMAKE_SOURCE_DIR}/qnx/remote_run_test.sh ${NAME})
   elseif(CMAKE_SYSTEM_NAME STREQUAL "Linux")
@@ -559,6 +559,7 @@ if((ENABLE_LOG) AND (CMAKE_SYSTEM_NAME STREQUAL "Linux"))
 
     set (install-3rd_party_logger_var "install-3rd_party_logger")
   endif()
+  set(LOG4CXX_LOGGER ON)
 endif()
 
 if (HMIADAPTER STREQUAL "dbus")
