+++ src/components/qt_hmi/qml_plugins/dbus_adapter/CMakeLists.txt
@@ -131,7 +131,7 @@ elseif (${qt_version} MATCHES "5.1.0")
 endif ()
 
 target_link_libraries(${target} log4cxx -L${LOG4CXX_LIBS_DIRECTORY})
-target_link_libraries(${target} Utils)
+target_link_libraries(${target} utils)
 
 add_custom_target(copy_library_${target} ALL
   COMMAND ${CMAKE_COMMAND} -E copy_if_different
