+++ src/components/utils/src/sql_qt_wrapper/sql_query.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, Ford Motor Company
+ * Copyright (c) 2016, Ford Motor Company
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,7 +30,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "sql_qt_wrapper/sql_query.h"
+#include "utils/sql_qt_wrapper/sql_query.h"
 
 #include <QSqlDatabase>
 #include <QSqlError>
