+++ src/components/utils/test/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright (c) 2015, Ford Motor Company
+# Copyright (c) 2016, Ford Motor Company
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -28,50 +28,37 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-set(UTILS_TEST_DIR ${COMPONENTS_DIR}/utils/test)
-
-find_package(SDLSqlite3 REQUIRED)
+include(${CMAKE_SOURCE_DIR}/tools/cmake/helpers/platform.cmake)
+include(${CMAKE_SOURCE_DIR}/tools/cmake/helpers/sources.cmake)
 
 include_directories (
-  ${JSONCPP_INCLUDE_DIRECTORY}
-  ${GMOCK_INCLUDE_DIRECTORY}
   ${COMPONENTS_DIR}/utils/include
   ${COMPONENTS_DIR}/policy/include
   ${COMPONENTS_DIR}/rpc_base/include
   ${COMPONENTS_DIR}/utils/test/include
-  ${SQLITE3_INCLUDE_DIRECTORY}
+  ${GMOCK_INCLUDE_DIRECTORY}
 )
 
-set(testSources
-  messagemeter_test.cc
-  file_system_test.cc
-  date_time_test.cc
-  thread_validator_test.cc
-  bitstream_test.cc
-  prioritized_queue_test.cc
-  data_accessor_test.cc
-  stl_utils_test.cc
-  async_runner_test.cc
-  shared_ptr_test.cc
-  scope_guard_test.cc
-  atomic_object_test.cc
-  message_loop_thread_test.cc
-  custom_string_test.cc
-  timer_test.cc
-  lock_test.cc
-  thread_test.cc
+set(EXCLUDE_PATHS
+  qdb_wrapper
+  test_generator
 )
 
-if (ENABLE_LOG)
-  list(APPEND testSources auto_trace_test.cc)
+if(NOT ENABLE_LOG)
+  list(APPEND EXCLUDE_PATHS
+    auto_trace_test.cc
+  )
 endif()
 
-if (BUILD_BACKTRACE_SUPPORT)
-  list(APPEND testSources back_trace_test.cc)
+if(NOT BUILD_BACKTRACE_SUPPORT)
+  list(APPEND EXCLUDE_PATHS
+    back_trace_test.cc
+  )
 endif()
 
-if (${CMAKE_SYSTEM_NAME} MATCHES "Linux")
-  list (APPEND testSources
+get_os(OS)
+if(NOT ${OS} STREQUAL "posix")
+  list(APPEND EXCLUDE_PATHS
     system_test.cc
     conditional_variable_test.cc
     message_queue_test.cc
@@ -79,38 +66,31 @@ if (${CMAKE_SYSTEM_NAME} MATCHES "Linux")
     rwlock_posix_test.cc
     resource_usage_test.cc
   )
-
 endif()
 
-set(testLibraries
+get_sdk(SDK)
+if(SDK)
+  if(${SDK} STREQUAL "qt")
+    list(APPEND EXCLUDE_PATHS
+      ${CMAKE_CURRENT_SOURCE_DIR}/sqlite_wrapper
+      ${CMAKE_CURRENT_SOURCE_DIR}/generated_code_with_sqlite_test.cc
+    )
+  endif()
+endif()
+
+collect_sources(SOURCES "${CMAKE_CURRENT_SOURCE_DIR}" "${EXCLUDE_PATHS}")
+
+set(LIBRARIES
   gmock
-  Utils
-  Policy
+  utils
+  Policy_static
   config_profile
 )
 
-if (CMAKE_SYSTEM_NAME STREQUAL "QNX")
-  # --- Tests for QDB Wrapper
-  list (APPEND testSources
-    ${UTILS_TEST_DIR}/qdb_wrapper/sql_database_test.cc
-    ${UTILS_TEST_DIR}/qdb_wrapper/sql_query_test.cc
-  )
-  file(COPY qdbserver.sh DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
-  file(COPY test-qdb.ini DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
-  file(COPY policy.sql DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
-elseif(NOT QT_PORT)
-  # --- Tests for SQLite Wrapper
-  list (APPEND testSources
-    ${UTILS_TEST_DIR}/sqlite_wrapper/sql_database_test.cc
-    ${UTILS_TEST_DIR}/sqlite_wrapper/sql_query_test.cc
-    ${UTILS_TEST_DIR}/generated_code_with_sqlite_test.cc
-  )
-endif()
+create_test(utils_test "${SOURCES}" "${LIBRARIES}")
 
 file(COPY testscript.sh DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
 file(COPY log4cxx.properties DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
 file(COPY smartDeviceLink.ini DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
-
-create_test("utils_test" "${testSources}" "${testLibraries}")
 
 add_subdirectory(test_generator)
