+++ tools/cmake/modules/FindSDLOpenSSL.cmake
@@ -14,5 +14,5 @@ elseif(${OS} STREQUAL "win")
     set(OPENSSL_DIR ${CMAKE_SOURCE_DIR}/build/openssl_win_x64)
   endif()
   set(OPENSSL_INCLUDE_DIRECTORY ${OPENSSL_DIR}/include)
-  set(OPENSSL_LIBRARIES ${OPENSSL_DIR}/lib/libeay32.lib ${OPENSSL_DIR}/lib/ssleay32.lib ws2_32)
+  set(OPENSSL_LIBRARIES ${OPENSSL_DIR}/lib/libeay32.lib ${OPENSSL_DIR}/lib/ssleay32.lib)
 endif()
